reorg and started tests
