+++ app/src/main/java/mozilla/lockbox/adapter/ItemListAdapter.kt
@@ -25,7 +25,8 @@ class ItemListAdapter : RecyclerView.Adapter<ItemViewHolder>() {
     private val compositeDisposable = CompositeDisposable()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ItemViewHolder {
-        val view = LayoutInflater.from(parent.context).inflate(R.layout.list_cell_item, parent, false)
+        val inflater = LayoutInflater.from(parent.context)
+        val view = inflater.inflate(R.layout.list_cell_item, parent, false)
 
         val viewHolder = ItemViewHolder(view)
         view
