+++ app/src/main/res/layout/list_cell_no_matching.xml
@@ -1,0 +1,48 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+  ~ This Source Code Form is subject to the terms of the Mozilla Public
+  ~ License, v. 2.0. If a copy of the MPL was not distributed with this
+  ~ file, You can obtain one at http://mozilla.org/MPL/2.0/.
+  -->
+
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent">
+
+    <TextView
+        android:layout_width="158dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:textAlignment="center"
+        android:textSize="17sp"
+        android:fontFamily="sans-serif"
+        android:textStyle="normal"
+        android:textColor="#0c0c0d"
+        android:lineSpacingExtra="8.5sp"
+        android:text="@string/no_matching_entries" />
+
+    <TextView
+        android:layout_width="280dp"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center_horizontal"
+        android:textAlignment="center"
+        android:textSize="13sp"
+        android:fontFamily="sans-serif"
+        android:textStyle="normal"
+        android:textColor="#0c0c0d"
+        android:lineSpacingExtra="6.5sp"
+        android:text="@string/save_firefox_disclaimer" />
+
+    <Button
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_gravity="center"
+        android:textSize="12sp"
+        android:fontFamily="sans-serif"
+        android:textStyle="bold"
+        android:textColor="#0060df"
+        android:letterSpacing="0.03"
+        android:lineSpacingExtra="7.5sp"
+        android:text="@string/learn_more" />
+
+</LinearLayout>
