+++ app/src/test/java/mozilla/lockbox/extensions/ServerPasswordExtensionText.kt
@@ -1,0 +1,11 @@
+/*
+ * This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at http://mozilla.org/MPL/2.0/.
+ */
+
+package mozilla.lockbox.extensions
+
+import org.junit.Assert.*
+
+class ServerPasswordExtensionText
